# specs/005-api-webhooks/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: EmailKit Public API
  description: |
    EmailKit API for email verification, webhook management, and credit balance queries.

    ## Authentication
    All endpoints (except `/health`) require Bearer token authentication using an API key.
    ```
    Authorization: Bearer ek_YOUR_API_KEY
    ```

    ## Rate Limiting
    Rate limits are enforced per-user based on subscription tier. All responses include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests
    - `X-RateLimit-Reset`: Unix timestamp when window resets

    ## Idempotency
    POST requests require an `Idempotency-Key` header (UUID v4). Duplicate requests
    within 24 hours return the cached response.

    ## Test Mode
    API keys with `ek_test_` prefix return mock responses without consuming credits.
  version: '1.0.0'
  contact:
    name: EmailKit Support
    email: support@emailkit.io
servers:
  - url: https://api.emailkit.io/api/v1
    description: Production API
security:
  - BearerAuth: []
...
paths:
...
  /webhooks:
    get:
      operationId: listWebhooks
      tags: [Webhooks]
      summary: List webhooks
      description: List all webhooks for the authenticated user.
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: after
          in: query
          schema:
            type: string
          description: Cursor for pagination
      responses:
        '200':
          description: List of webhooks
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Webhook'
                  hasMore:
                    type: boolean
                  nextCursor:
                    type: string
                    nullable: true
                  requestId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      operationId: createWebhook
      tags: [Webhooks]
      summary: Create webhook
      description: |
        Create a new webhook endpoint. The URL must be HTTPS and will be
        validated with a test ping before activation.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                  description: HTTPS endpoint URL
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/WebhookEventType'
                  minItems: 1
                  description: Event types to subscribe to
                payloadMode:
                  type: string
                  enum: [full, summary]
                  default: full
                  description: Level of detail in webhook payloads
      responses:
        '201':
          description: Webhook created
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Webhook'
                  - type: object
                    properties:
                      secret:
                        type: string
                        description: |
                          Signing secret (shown once). Use to verify webhook signatures.
                          Format: whsec_<random>
        '400':
          description: URL validation failed (test ping did not receive 2xx)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Webhook limit reached (max 10)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          $ref: '#/components/responses/ValidationError'
  /webhooks/{webhookId}:
    get:
      operationId: getWebhook
      tags: [Webhooks]
      summary: Get webhook
      description: Retrieve a specific webhook by ID.
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Webhook details
          headers:
            X-Request-ID:
              $ref: '#/components/headers/X-Request-ID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      operationId: updateWebhook
      tags: [Webhooks]
      summary: Update webhook
      description: Update webhook configuration. Can update events, URL, payload mode, or reactivate paused webhook.
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    $ref: '#/components/schemas/WebhookEventType'
                payloadMode:
                  type: string
                  enum: [full, summary]
                status:
                  type: string
                  enum: [active]
                  description: Set to 'active' to reactivate a paused webhook
      responses:
        '200':
          description: Webhook updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: URL validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteWebhook
      tags: [Webhooks]
      summary: Delete webhook
      description: Permanently delete a webhook.
      parameters:
        - name: webhookId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Webhook deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Use your API key as the bearer token.
        ```
        Authorization: Bearer ek_YOUR_API_KEY
        ```
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: UUID v4 for request deduplication. Required for all POST requests.
  headers:
    X-Request-ID:
      schema:
        type: string
      description: Unique request identifier for debugging
...
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        url:
          type: string
          format: uri
        events:
          type: array
          items:
            $ref: '#/components/schemas/WebhookEventType'
        payloadMode:
          type: string
          enum: [full, summary]
        status:
          type: string
          enum: [active, failing, paused]
        failureCount:
          type: integer
        lastDeliveryAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    WebhookEventType:
      type: string
      enum:
        - verification.completed
        - bulk.completed
        - bulk.failed
        - credits.low
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          nullable: true
          description: Additional error details
        requestId:
          type: string
  responses:
    Unauthorized:
      description: Invalid or missing API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Invalid or expired API key
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: The requested resource was not found
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  fields:
                    type: object
                    additionalProperties:
                      type: string
          example:
            error: validation_error
            message: Request validation failed
            fields:
              email: Invalid email format
...